# Conflicts:
#	WMS/App/App/WmsApplication.cs
#	WMS/AutomationServiceMock/AutomationService.cs
#	WMS/AutomationServiceMock/Program.cs
#	WMS/Data/WebAPI.Contracts/DataServiceFactory.cs
#	WMS/Data/WebAPI.Contracts/Ferretto.WMS.Data.WebAPI.Contracts.csproj
#	WMS/Data/WebAPI.Contracts/ServiceContracts.autogen.cs
#	WMS/Data/WebAPI/Ferretto.WMS.Data.WebAPI.csproj
#	WMS/Data/WebAPI/Startup.cs
#	WMS/WMS.sln

+++ Common/DataModels/ItemListRow.cs
@@ -14,7 +14,7 @@ namespace Ferretto.Common.DataModels
 
         public DateTime CreationDate { get; set; }
 
-        public int DispatchedQuantity { get; set; }
+        public double DispatchedQuantity { get; set; }
 
         public int Id { get; set; }
 
@@ -46,7 +46,7 @@ namespace Ferretto.Common.DataModels
 
         public string RegistrationNumber { get; set; }
 
-        public int RequestedQuantity { get; set; }
+        public double RequestedQuantity { get; set; }
 
         public IEnumerable<SchedulerRequest> SchedulerRequests { get; set; }
